requirements.txt
absl-py==2.1.0
albucore==0.0.17
albumentations==1.4.18
annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
contextlib2==21.6.0
contourpy==1.1.1
cycler==0.12.1
docopt==0.6.2
einops==0.8.0
eval_type_backport==0.2.2
filelock==3.16.1
fonttools==4.55.1
fsspec==2024.12.0
huggingface-hub==0.28.1
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
loralib==0.1.2
matplotlib==3.7.5
MedPy==0.5.2
ml_collections==0.1.1
monai==1.3.2
networkx==3.1
numpy==1.24.4
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==24.2
pandas==2.0.3
patsy==1.0.1
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
pykwalify==1.8.0
pyparsing==3.1.4
pyradiomics==3.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyWavelets==1.4.1
PyYAML==6.0.2
requests==2.32.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.8
safetensors==0.5.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.2
SimpleITK==2.4.0
six==1.16.0
some-package==0.1
statsmodels==0.14.1
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.14
torch==1.13.0+cu116
torchaudio==0.13.0+cu116
torchvision==0.14.0+cu116
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
yacs==0.1.8
zipp==3.20.2
